successfully added the functionalities manualy that allows redux keep tract of react-router history
Before this addition, i noticed there is always an error anytime i try to go back to a previous page in my app, so i found out after debugging that the readProps attribute of the history.location props is missing, it is the readProps attribute that allows redux to fetch data based on the current page. To rectify this i created an History array in the global that monitor page changes if a page change request is made, the current page data is transferred into the History array. When back is clicked now the last page data is pop from the global history state and fed to the react-router history object. One more thing ones the whole react-router stack has been cleared, the global history object is set back to an empty array